Move edge_cases and their  ddl to integration level fixtures
